==> 5_Post_Processing_Testing_Locations.ipynb <==
# Post-process: Smoothing and Reclassify Classes

## Background
In real case studies machine learning may not predict desired classification maps due to factors including limited training data and limited performance of the model for prediction. Therefore, post-processing is widely applied to the predicted classification results, based on assumptions or existing knowlegde of the ground truth. Commonly applied post-processing includes manual editing, filtering, reclassification and class merging, etc.  

## Description
In this notebook we will apply post-processing to the land cover maps we produced through the previous notebook. We'll also conduct a median filtering to reduce the 'salt and pepper' effect resulted from pixel-based classification. We will then use external layers that contain reliable information on certain classes and/or have higher-spatial resolution to reclassify classes that may be misclassified by the random forest classifier. These external layers have been prepared and uploaded into 'Data/' folder. This notebook will demonstrate how to do implement these post-processings and visualise the comparison before and after the post-processing. The steps are as follows:
1. Load the external layers and land cover maps produced at the testing locations
2. Majority filtering of the maps to reduce salt-and-pepper effects
3. Apply customised reclassification rules using prepared external layers
4. Plot the land cover maps before and after post-processing steps applied
5. Save the results to disk as COGs

To run this analysis, run all the cells in the notebook, starting with the "Load packages" cell.

## Load Packages

%matplotlib inline
import numpy as np
import geopandas as gpd
import xarray as xr
import rioxarray
import datacube
from datacube.utils.cog import write_cog
from deafrica_tools.spatial import xr_rasterize
from deafrica_tools.plotting import rgb
from skimage.morphology import binary_dilation,disk
from skimage.filters.rank import modal
from skimage.segmentation import expand_labels
from odc.algo import xr_reproject
import matplotlib.pyplot as plt
from matplotlib.colors import ListedColormap,BoundaryNorm
from matplotlib.patches import Patch

## Analysis parameters
* `prediction_maps_path`: A list of file paths and names of the classification maps produced in the previous notebook.
* `rgb_images_path`: A list of file paths and names of the true colour images at the prediction locations exported in the previous notebook.
* `dict_map`: A dictionary map of class names corresponding to pixel values.
* `output_crs`: Coordinate reference system for output raster files.

prediction_maps_path=['Results/Land_cover_prediction_Kicukiro_location_'+str(i)+'.tif' for i in range(3)] # list of prediction map files
rgb_images_path=['Results/S2_RGB_image_Kicukiro_location_'+str(i)+'.tif' for i in range(3)] # list of rgb images at the prediction locations
dict_map={'Forest':1,'Grassland':5,'Shrubland':7,'Perennial Cropland':9,'Annual Cropland':10,
          'Wetland':11,'Water Body':12,'Urban Settlement':13,'Bare Land':14} # a dictionary of pixel value for each class
output_crs='epsg:32735' # WGS84/UTM Zone 35S

## External Layers
A few external layers for Rwanda were sourced and prepared in the 'Data/' folder, which are helpful to add information on specific classes. The external layers include:  
* `river_network_shp`: OSM river network shapefile. The OSM layers were sourced from the [Humanitarian OpenStreetMap Team (HOT)](https://data.humdata.org/organization/hot) website.
* `road_network_shp`: OSM road network shapefile.
* `google_building_raster`: A rasterised layer of [Google Open Building](https://developers.google.com/earth-engine/datasets/catalog/GOOGLE_Research_open-buildings_v2_polygons) polygons, which consist of outlines of buildings derived from high-resolution 50 cm satellite imagery. As there are many polygons in the original vector layer, we rasterised the layer to 10 m resolution to reduce disk storage and memory required for processing.  
* `wsf2019_raster`: 2019 [World Settlement Footprint (WSF)](https://gee-community-catalog.org/projects/wsf/) layer, a 10m resolution binary mask outlining the extent of human settlements globally derived by means of 2019 multitemporal Sentinel-1 and Sentinel-2 imagery.  
* `hand_raster`: Hydrologically adjusted elevations, i.e. Height Above the Nearest Drainage (hand) derived from the [MERIT Hydro dataset](https://developers.google.com/earth-engine/datasets/catalog/MERIT_Hydro_v1_0_1#description).
* `parks_json`: Boundaries of [national parks or reserve](https://geodata.rw/server/rest/services/basemap/National_Parks/FeatureServer/0) declared or registered by the government in geojson format.
* `wetlands_json`: Boundaries of [declaired swamps borders](https://geodata.rw/server/rest/services/basemap/Wetland/FeatureServer/0) in geojson format.

> Note: In this notebook we have made the data prepared for you to run through the demonstration. If you would like to apply it to your own project, you may need to spend some time sourcing the datasets and do some pre-processing if needed, e.g. clipping to your study area, filtering, rasterisation or vectorisation. Alternatively you can revise this notebook depending on the file format of your data.

river_network_shp='Data/hotosm_rwa_waterways_lines_filtered.shp' # OSM river network data
road_network_shp='Data/hotosm_rwa_roads_lines_filtered.shp' # OSM road network data
google_building_raster='Data/GoogleBuildingLayer_Rwanda_reprojected_rasterised.tif' # rasterised google bulding layer
wsf2019_raster='Data/WSF2019_v1_Rwanda_clipped.tif' # 2019 World Settlement Footprint layer
hand_raster='Data/hand_Rwanda.tif' # Hydrologically adjusted elevations, i.e. height above the nearest drainage (hand)
parks_json='Data/National Parks.geojson' # National parks layer
wetlands_json='Data/Wetland.geojson' # wetlands layer

## Load land cover maps and external layers
First let's load the land cover maps and true colour images generated from the previous notebook:

# import land cover map of 2021 and reproject
prediction_maps=[]
rgb_images=[]
for i in range(0, len(prediction_maps_path)):
    lc_map=rioxarray.open_rasterio(prediction_maps_path[i]).astype(np.uint8).squeeze()
    prediction_maps.append(lc_map)
    
    rgb_image=rioxarray.open_rasterio(rgb_images_path[i])
    rgb_images.append(rgb_image)

We then load other layers. The OSM road network layer contains multi-lines with various surface attributes. We'll select some major road types and buffer them by 10 metres. Similarly we load and select major waterways from the OSM river network layer.

road_network=gpd.read_file(road_network_shp).to_crs(output_crs) 
road_network=road_network.loc[road_network['surface'].isin(['asphalt', 'paved', 'compacted', 'cobblestone', 
                                                             'concrete', 'metal', 'paving_stones', 
                                                             'paving_stones:30'])] # select road network by attributes
road_network.geometry=road_network.geometry.buffer(10) # buffer the road network by 10m

river_network=gpd.read_file(river_network_shp).to_crs(output_crs) # import OSM river network data and reproject
river_network=river_network.loc[river_network['waterway'].isin(['canal','river'])] # select river network by attribute

We now load the Google buildings, WSF 2019 and 'hand' rasters:

google_buildings=xr.open_dataset(google_building_raster,engine="rasterio").squeeze() # import google bulding layer
hand=xr.open_dataset(hand_raster,engine="rasterio").squeeze() # import hand layer
wsf2019=xr.open_dataset(wsf2019_raster,engine="rasterio").squeeze().astype(np.uint8)

Finally let's load the national parks and wetlands layers, and select the national parks with no agricultural activities:

parks=gpd.read_file(parks_json).to_crs(output_crs)
parks=parks.loc[parks['type'].isin(['National Park','Volcanoes National Park'])]
wetlands=gpd.read_file(wetlands_json).to_crs(output_crs)

## Morphological filtering
Now we start the post-processing by applying a majority filtering, a commonly applied step to reduce salt-and-pepper noise typical in pixel-based classification. To demonstrate each post-processing step we will process the first prediction map, then put the steps together in an iterative loop to process all prediction. The majority filtering is applied within each local window with a given footprint. Here we use a disk with radius of one pixel. It is advised that you adjust the footprint depending on your prediction results and desired effects.

i=0
lc_map=prediction_maps[i]
# convert to numpy array
np_lc_map=lc_map.squeeze().to_numpy()
# mode filtering for a smoother classification map
np_lc_map_postproc=modal(np_lc_map,footprint=disk(1),mask=np_lc_map!=0)

We can plot and compare the maps before and after filtering:

# reconstruct dataArray
lc_map_postproc=xr.DataArray(data=np_lc_map_postproc,dims=['y','x'],
                             coords={'y':lc_map.y.to_numpy(), 'x':lc_map.x.to_numpy()})
# display colour for each class value
colours = {1:'green', 5:'gold', 7:'chocolate',9:'violet',10:'pink',11:'cyan',12:'blue',13:'gray',14:'seashell'}

fig, axes = plt.subplots(1, 2, figsize=(24, 8))

# Plot classified image before filtering
prediction_values=np.unique(lc_map)
cmap=ListedColormap([colours[k] for k in prediction_values])
norm = BoundaryNorm(list(prediction_values)+[np.max(prediction_values)+1], cmap.N)
lc_map.plot.imshow(ax=axes[0], 
                   cmap=cmap,
                   norm=norm,
                   add_labels=True, 
                   add_colorbar=False,
                   interpolation='none')

# Plot classified image after filtering
lc_map_postproc.plot.imshow(ax=axes[1], 
                   cmap=cmap,
                   norm=norm,
                   add_labels=True, 
                   add_colorbar=False,
                   interpolation='none')
# Remove axis on right plot
axes[1].get_yaxis().set_visible(False)
# add colour legend
patches_list=[Patch(facecolor=colour) for colour in colours.values()]
axes[1].legend(patches_list, list(dict_map.keys()),loc='upper center', ncol =4, bbox_to_anchor=(0.5, -0.1))
# Add plot titles
axes[0].set_title('Classified Image Before Majority Filtering')
axes[1].set_title('Classified Image After Majority Filtering')

# make a copy of intermediate result
lc_map_filtered=lc_map_postproc.copy()

## Apply rules using external layers
Now let's reclassify some classes using the external layers. We will apply the following processing rules:  
* Merging Perennial Cropland and Annual Cropland as Cropland class;
* Assign Cropland pixels within national parks as Shrubland;
* Assign Urban Settlement pixels within national parks as surrounding classes;
* Assign Wetland pixels outside wetlands polygons as Shrubland;
* Assign Bare Land or Urban Settlement pixels within DE Africa 2019 cropland mask as Cropland;
* Assign Cropland pixels outside DE Africa 2019 cropland mask as Shrubland;
* Assign Water Body pixels as surrounding class if they don’t occur at bottom of watersheds and fall outside OSM river networks;
* Assign pixels overlapping OSM river network as Water Body;
* Assign pixels overlapping google building polygons, WSF 2019 or OSM road network (10 m buffered) layers as Urban Settlement.

> Note: The reclassification rules applied in this notebook are very customised and may not apply to other countries or datasets. What rules to apply depend on various factors including your preliminary knowledge on the study area, the reliability of the external layers and your classes of most interest/priority. Therefore, you may need to revise or refine the processing rules applied here to fit your own application needs.

Before applying reclassification using other layers, one thing to note for raster-based calculation is to make sure all rasters are in the same spatial reference and align with each other. Here we extract the `geobox` of the land cover map, which defines the location and resolution of the grid data including spatial reference. We will use it to reproject other layers to be aligned with the land cover map.

ds_geobox=lc_map.geobox

First, we merge Perennial Cropland and Annual Cropland as a single class Cropland:

np_lc_map_postproc[np_lc_map_postproc==dict_map['Perennial Cropland']]=dict_map['Annual Cropland']

As there is unlikely agricultural activities within national parks, the Cropland pixels within national parks are likely to be misclassified. After checking on the results, we believe they are most likely to be Shrubland, so we assign these misclassified pixels as Shrubland. Similarly we assign Urban Settlement pixels within national parks as surrounding classes:

# rasterise national parks layer
parks_mask=xr_rasterize(gdf=parks,da=lc_map.squeeze(),transform=ds_geobox.transform,crs=output_crs)
np_parks_mask=parks_mask.squeeze().to_numpy()
# apply the rule
np_lc_map_postproc[(np_lc_map_postproc==dict_map['Annual Cropland'])&(np_parks_mask==1)]=dict_map['Shrubland']

# make a copy of land cover array
lc_copy=np_lc_map_postproc.copy()
# assign the regions as background (0)
lc_copy[(np_lc_map_postproc==dict_map['Urban Settlement'])&(np_parks_mask==1)]=0
# expand surrounding classes of background pixels
lc_copy_closed=expand_labels(lc_copy,distance=10000)
# identify filled/changed areas
mask=(lc_copy!=lc_copy_closed)
# copy the filled/changed pixels
np_lc_map_postproc[mask]=lc_copy_closed[mask]

We believe that the wetland layer includes all wetlands, so we assign Wetland pixels outside the wetland polygons as Shrubland:

wetlands_mask=xr_rasterize(gdf=wetlands,da=lc_map.squeeze(),transform=ds_geobox.transform,crs=output_crs)
np_wetlands_mask=wetlands_mask.squeeze().to_numpy()
np_lc_map_postproc[(np_lc_map_postproc==dict_map['Wetland'])&(np_wetlands_mask!=1)]=dict_map['Shrubland']

We observed that there are more Bare Land and Urban Settlement pixels than expected, and a lot of them appear to be croplands, so we use the [DE Africa 2019 cropland mask](https://docs.digitalearthafrica.org/en/latest/data_specs/Cropland_extent_specs.html) and assign the misclassified pixels as Cropland. Besides, we believe that the crop mask is comprehensive, so cropland pixels outside the mask are likely to have been misclassified from other classes. In this example they are mostly likely to be Shrubland, so we assign these cropland pixels as Shrubland:

# load DE Africa crop mask 2019
dc = datacube.Datacube(app='cropland_extent')
x_min,y_min,x_max,y_max=ds_geobox.extent.boundingbox
query = {
    'time': ('2019'),
    'x': (x_min,x_max),
    'y': (y_min,y_max),
    'resolution':(-10, 10),
    'crs':output_crs,
    'output_crs': output_crs,
}
# now load the crop-mask using the query
cm = dc.load(product='crop_mask',**query).squeeze()
np_crop_mask=cm['mask'].to_numpy()
# apply the rules
np_lc_map_postproc[((np_lc_map_postproc==dict_map['Bare Land'])|(np_lc_map_postproc==dict_map['Urban Settlement']))
&(np_crop_mask==1)]=dict_map['Annual Cropland']
np_lc_map_postproc[(np_lc_map_postproc==dict_map['Annual Cropland'])&(np_crop_mask!=1)]=dict_map['Shrubland']

To include more Water Body pixels that are misclassified to other classes, we use the OSM river network layer and assign pixels overlapping the layer as Water Body. Meanwhile, to make sure water only occurs at bottom of watersheds, i.e. height above nearest drainage below 45m, or falls within OSM river networks, we assign Water Body pixels outside these areas as surrounding classes:

# reproject hand layer
hand=xr_reproject(hand, ds_geobox, resampling="average")
# convert to numpy array
np_hand=hand.to_array().squeeze().to_numpy()
# rasterise river network layer
river_network_mask=xr_rasterize(gdf=river_network,
                                  da=lc_map.squeeze(),
                                  transform=ds_geobox.transform,
                                  crs=output_crs)
# convert to numpy array
np_river_network_mask=river_network_mask.to_numpy()
# apply the rule
np_lc_map_postproc[np_river_network_mask==1]=dict_map['Water Body']

# make a copy of land cover array
lc_copy=np_lc_map_postproc.copy()
# assign the regions as background (0)
lc_copy[(np_lc_map_postproc==dict_map['Water Body'])&(np_hand>45)&(np_river_network_mask!=1)]=0
# expand surrounding classes of background pixels
lc_copy_closed=expand_labels(lc_copy,distance=10000)
# identify filled/changed areas
mask=(lc_copy!=lc_copy_closed)
# copy the filled/changed pixels
np_lc_map_postproc[mask]=lc_copy_closed[mask] 

Finally, we assign pixels overlapping google building polygons, WSF 2019 mask or OSM road network (10m buffered) as Urban Settlement:

# reproject google buildings raster
google_buildings_mask=xr_reproject(google_buildings, ds_geobox, resampling="average")
# convert to numpy array
np_google_buildings_mask=google_buildings_mask.to_array().squeeze().to_numpy()
# reproject WSF 2019 layer
wsf2019=xr_reproject(wsf2019, ds_geobox, resampling="nearest")
# convert to numpy array
np_wsf2019=wsf2019.to_array().squeeze().to_numpy()
# rasterise road network layer
road_network_mask=xr_rasterize(gdf=road_network,
                              da=lc_map.squeeze(),
                              transform=ds_geobox.transform,
                              crs=output_crs)
# convert to numpy
np_road_network_mask=road_network_mask.to_numpy()
# apply rule
np_lc_map_postproc[(np_google_buildings_mask==1)|(np_wsf2019==255)|(np_road_network_mask==1)]=dict_map['Urban Settlement']

We can plot the maps to see a comparison before and after applying the rules using the external layers:

fig, axes = plt.subplots(1, 2, figsize=(24, 8))

# set color legends and color maps parameters
prediction_values=np.unique(lc_map_filtered)
cmap=ListedColormap([colours[k] for k in prediction_values])
norm = BoundaryNorm(list(prediction_values)+[np.max(prediction_values)+1], cmap.N)

dict_map_merged={'Forest':1,'Grassland':5,'Shrubland':7,'Cropland':10,
          'Wetland':11,'Water Body':12,'Urban Settlement':13,'Bare Land':14}
colours_merged={1:'green', 5:'gold', 7:'chocolate',10:'pink',11:'cyan',12:'blue',13:'gray',14:'seashell'}
patches_list_merged=[Patch(facecolor=colour) for colour in colours_merged.values()]
prediction_values_merged=np.unique(lc_map_postproc)
cmap_merged=ListedColormap([colours_merged[k] for k in prediction_values_merged])
norm_merged = BoundaryNorm(list(prediction_values_merged)+[np.max(prediction_values_merged)+1], cmap_merged.N)

# Plot classified image before applying rules
lc_map_filtered.plot.imshow(ax=axes[0], 
                   cmap=cmap,
                   norm=norm,
                   add_labels=True, 
                   add_colorbar=False,
                   interpolation='none')

# Plot classified image after applying rules
# reconstruct dataArray
lc_map_postproc=xr.DataArray(data=np_lc_map_postproc,dims=['y','x'],
                             coords={'y':lc_map.y.to_numpy(), 'x':lc_map.x.to_numpy()})
lc_map_postproc.plot.imshow(ax=axes[1], 
                   cmap=cmap_merged,
                   norm=norm_merged,
                   add_labels=True, 
                   add_colorbar=False,
                   interpolation='none')
# Remove axis on right plot
axes[1].get_yaxis().set_visible(False)
# add colour legends
axes[0].legend(patches_list, list(dict_map.keys()),
    loc='upper center', ncol =4, bbox_to_anchor=(0.5, -0.1))
axes[1].legend(patches_list_merged, list(dict_map_merged.keys()),
    loc='upper center', ncol =4, bbox_to_anchor=(0.5, -0.1))
# Add plot titles
axes[0].set_title('Classified Image Before Applying Rules')
axes[1].set_title('Classified Image After Applying Rules')

## Processing and visualisation for all testing locations
Now let's process all three locations by simply copying and putting together all the post-processing steps, and iterating through all three locations:

prediction_maps_postproc=[] # post-processed results
# loop through all testing locations
for i in range(0,len(prediction_maps)):
    lc_map=prediction_maps[i]
    # convert to numpy array
    np_lc_map=lc_map.squeeze().to_numpy()

    # --------------------mode filtering for a smoother classification map----------------------#
    np_lc_map_postproc=modal(np_lc_map,footprint=disk(1),mask=np_lc_map!=0)

    # get geobox
    ds_geobox=lc_map.geobox

    # merging two cropland classes
    np_lc_map_postproc[np_lc_map_postproc==dict_map['Perennial Cropland']]=dict_map['Annual Cropland']

    #--------------------reassign Cropland within national parks as Shrubland------------------------#
    # rasterise national parks layer
    parks_mask=xr_rasterize(gdf=parks,da=lc_map.squeeze(),transform=ds_geobox.transform,crs=output_crs)
    np_parks_mask=parks_mask.squeeze().to_numpy()
    # apply the rule
    np_lc_map_postproc[(np_lc_map_postproc==dict_map['Annual Cropland'])&(np_parks_mask==1)]=dict_map['Shrubland']

    #--------------------reassign Urban Settlement within national parks as surrounding classes------------------------#
    # make a copy of land cover array
    lc_copy=np_lc_map_postproc.copy()
    # assign the regions as background (0)
    lc_copy[(np_lc_map_postproc==dict_map['Urban Settlement'])&(np_parks_mask==1)]=0
    # expand surrounding classes of background pixels
    lc_copy_closed=expand_labels(lc_copy,distance=10000)
    # identify filled/changed areas
    mask=(lc_copy!=lc_copy_closed)
    # copy the filled/changed pixels
    np_lc_map_postproc[mask]=lc_copy_closed[mask]

    #--------------------reassign Wetland outside wetland polygons as Shrubland------------------------#
    wetlands_mask=xr_rasterize(gdf=wetlands,da=lc_map.squeeze(),transform=ds_geobox.transform,crs=output_crs)
    np_wetlands_mask=wetlands_mask.squeeze().to_numpy()
    np_lc_map_postproc[(np_lc_map_postproc==dict_map['Wetland'])&(np_wetlands_mask!=1)]=dict_map['Shrubland']

    #------reassign Bare Land and Urban Settlement within DE Africa crop mask as Cropland; Cropland outside the mask as Shrubland------#
    # load DE Africa crop mask 2019
    dc = datacube.Datacube(app='cropland_extent')
    # extract bounding box
    x_min,y_min,x_max,y_max=ds_geobox.extent.boundingbox
    query = {
        'time': ('2019'),
        'x': (x_min,x_max),
        'y': (y_min,y_max),
        'resolution':(-10, 10),
        'crs':output_crs,
        'output_crs': output_crs,
    }
    # now load the crop-mask using the query
    cm = dc.load(product='crop_mask',**query).squeeze()
    np_crop_mask=cm['mask'].to_numpy()
    # apply the rules
    np_lc_map_postproc[((np_lc_map_postproc==dict_map['Bare Land'])|(np_lc_map_postproc==dict_map['Urban Settlement']))
    &(np_crop_mask==1)]=dict_map['Annual Cropland']
    np_lc_map_postproc[(np_lc_map_postproc==dict_map['Annual Cropland'])&(np_crop_mask!=1)]=dict_map['Shrubland']

    #--------------------assign pixels within OSM river network as Water Body------------------------#
    # reproject hand layer
    hand=xr_reproject(hand, ds_geobox, resampling="average")
    # convert to numpy array
    np_hand=hand.to_array().squeeze().to_numpy()
    # rasterise river network layer
    river_network_mask=xr_rasterize(gdf=river_network,
                                    da=lc_map.squeeze(),
                                    transform=ds_geobox.transform,
                                    crs=output_crs)
    # convert to numpy array
    np_river_network_mask=river_network_mask.to_numpy()
    # apply the rule
    np_lc_map_postproc[np_river_network_mask==1]=dict_map['Water Body']

    #--------------------make sure water body only present at bottom of watershed or within OSM river network-----------#
    # make a copy of land cover array
    lc_copy=np_lc_map_postproc.copy()
    # assign the regions as background (0)
    lc_copy[(np_lc_map_postproc==dict_map['Water Body'])&(np_hand>45)&(np_river_network_mask!=1)]=0
    # expand surrounding classes of background pixels
    lc_copy_closed=expand_labels(lc_copy,distance=10000)
    # identify filled/changed areas
    mask=(lc_copy!=lc_copy_closed)
    # copy the filled/changed pixels
    np_lc_map_postproc[mask]=lc_copy_closed[mask]

    #--------------------assign pixels overlapping google building, WSF 2019 and OSM road network as Urban Settlement---------#
    # reproject google buildings raster
    google_buildings_mask=xr_reproject(google_buildings, ds_geobox, resampling="average")
    # convert to numpy array
    np_google_buildings_mask=google_buildings_mask.to_array().squeeze().to_numpy()
    # reproject WSF 2019 layer
    wsf2019=xr_reproject(wsf2019, ds_geobox, resampling="nearest")
    # convert to numpy array
    np_wsf2019=wsf2019.to_array().squeeze().to_numpy()
    # rasterise road network layer
    road_network_mask=xr_rasterize(gdf=road_network,
                                da=lc_map.squeeze(),
                                transform=ds_geobox.transform,
                                crs=output_crs)
    # convert to numpy
    np_road_network_mask=road_network_mask.to_numpy()
    # apply rule
    np_lc_map_postproc[(np_google_buildings_mask==1)|(np_wsf2019==255)|(np_road_network_mask==1)]=dict_map['Urban Settlement']

    #------------------------------reconstruct processed result as Xarray---------------------------------#
    # 
    lc_map_postproc=xr.DataArray(data=np_lc_map_postproc,dims=['y','x'],
                             coords={'y':lc_map.y.to_numpy(), 'x':lc_map.x.to_numpy()})
    # set spatial reference
    lc_map_postproc.rio.write_crs(output_crs, inplace=True)
    # append to list
    prediction_maps_postproc.append(lc_map_postproc)

We can compare all final post-processed results with initial predictions without post-processing, along with the satellite images for comparison:

for i in range(0, len(prediction_maps)):
    fig, axes = plt.subplots(1, 3, figsize=(24, 8))

    # Plot true colour image
    rgb(rgb_images[i].to_dataset(dim='band'),bands=[1,2,3],
        ax=axes[0], percentile_stretch=(0.01, 0.99))
    
    # set color legends and color maps parameters
    prediction_values=np.unique(prediction_maps[i])
    cmap=ListedColormap([colours[k] for k in prediction_values])
    norm = BoundaryNorm(list(prediction_values)+[np.max(prediction_values)+1], cmap.N)

    prediction_values_merged=np.unique(prediction_maps_postproc[i])
    cmap_merged=ListedColormap([colours_merged[k] for k in prediction_values_merged])
    norm_merged = BoundaryNorm(list(prediction_values_merged)+[np.max(prediction_values_merged)+1], cmap_merged.N)
    
    # Plot initial classified image
    prediction_maps[i].plot.imshow(ax=axes[1], 
                   cmap=cmap,
                   norm=norm,
                   add_labels=True, 
                   add_colorbar=False,
                   interpolation='none')
    
    # Plot post-processed classified image
    prediction_maps_postproc[i].plot.imshow(ax=axes[2], 
                   cmap=cmap_merged,
                   norm=norm_merged,
                   add_labels=True, 
                   add_colorbar=False,
                   interpolation='none')
                   
    # Remove axis on middle and right plot
    axes[1].get_yaxis().set_visible(False)
    axes[2].get_yaxis().set_visible(False)
    # add colour legends
    axes[1].legend(patches_list, list(dict_map.keys()),
        loc='upper center', ncol =4, bbox_to_anchor=(0.5, -0.1))
    axes[2].legend(patches_list_merged, list(dict_map_merged.keys()),
        loc='upper center', ncol =4, bbox_to_anchor=(0.5, -0.1))
    # Add plot titles
    axes[0].set_title('True Colour Geomedian')
    axes[1].set_title('Classified Image')
    axes[2].set_title('Classified Image - Postprocessed')

## Save as geotiff
We can now export our post-processed results to sandbox disk as Cloud-Optimised GeoTIFFs:

for i in range(0, len(prediction_maps_postproc)):
    write_cog(prediction_maps_postproc[i], 'Results/Land_cover_prediction_postprocessed_Kicukiro_location_'+str(i)+'.tif', overwrite=True)
